+++ freud/box.pyx
@@ -718,15 +718,7 @@ cdef class Box:
             pass
         return NotImplemented
 
-    def __mul__(arg1, arg2):
-        # Note Cython treats __mul__ and __rmul__ as one operation, so
-        # type checks are necessary.
-        if isinstance(arg1, freud.box.Box):
-            self = arg1
-            scale = arg2
-        else:
-            scale = arg1
-            self = arg2
+    def __mul__(self, scale):
         if scale > 0:
             return self.__class__(Lx=self.Lx*scale,
                                   Ly=self.Ly*scale,
@@ -735,6 +727,9 @@ cdef class Box:
                                   is2D=self.is2D)
         else:
             raise ValueError("Box can only be multiplied by positive values.")
+
+    def __rmul__(self, scale):
+        return self * scale
 
     def plot(self, title=None, ax=None, image=[0, 0, 0], *args, **kwargs):
         """Plot a :class:`~.box.Box` object.